+++ src/components/Organisms/Header/styles.tsx
@@ -1,0 +1,10 @@
+'use client';
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100px;
+`;
+
