--- src/components/Templates/Header/index.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import { Container, Title  } from './styles'
-
-const Header: React.FC = () => {
-  return (
-      <Container>
-        <Title>Header</Title>
-      </Container>
-  )
-}
-
-export default Header
