--- src/components/Templates/Header/styles.tsx
@@ -1,14 +1,0 @@
-'use client';
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  background-color: #f1f1f1;
-  padding: 16px;
-  border-radius: 4px;
-`;
-
-export const Title = styled.h1`
-  color: #333;
-  font-size: 24px;
-  margin-bottom: 16px;
-`;
